A FUNCTIONAL THING
score calculation refinement + bug fixing
refined loss condition
integration of shooty stuff with disease spread stuff
rudimentary playtime UI

+++ Deadline Dread/Assets/Scripts/Foam_Launcher.cs
@@ -23,7 +23,7 @@ public class Foam_Launcher : MonoBehaviour
     public float bulletCount;
     public int maxAmmo;
     public int ammoCount;
-    public int ammoRegen;
+    public float ammoRegen;
     private System.Random rFactor = new System.Random();
     public Sprite sprite;
     public bool fireMode;
